requirements.txt
asgiref==3.8.1
boto3==1.38.3
botocore==1.38.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.6
django-anymail==13.0
django-cors-headers==4.7.0
django-filter==24.3
django-js-asset==3.1.2
django-mptt==0.17.0
django-storages==1.14.5
django-taggit==6.1.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
easy-thumbnails==2.10
gunicorn==21.2.0
idna==3.10
jmespath==1.0.1
Markdown==3.7
oauthlib==3.2.2
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.12.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
